% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBARtabQC.R
\name{aggregateBarcodes}
\alias{aggregateBarcodes}
\title{aggregateBarcodes}
\usage{
aggregateBarcodes(counts)
}
\arguments{
\item{counts}{Dataframe with barcodes and UMI counts per cell}
}
\value{
Returns a data frame with one row per cell ID
}
\description{
Aggregate barcodes and barcode UMIs per cell
}
