% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateBarcodeReads.R
\name{annotateBarcodeReads}
\alias{annotateBarcodeReads}
\title{Annotate Barcode Reads}
\usage{
annotateBarcodeReads(all.cells, barcoded.cells)
}
\arguments{
\item{all.cells}{list of all cells in the experiment to be annotated}

\item{barcoded.cells}{dataframe a dataframe with cell id and barcode id as columns}
}
\value{
Returns a data-frame containing the the 10X cell ID and the lintraceR DNA barcode ID.
}
\description{
Matches cells in a single cell experiment to detected DNA barcodes.
Dataframe returned will have all cells matched to a barcode
If there is no barcode matchable to a cell "not.detected" is returned
For cells that have multiple detected barcodes each barcode is returned separated by ';'
}
