% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDivIndexes.R
\name{calcDivIndexes}
\alias{calcDivIndexes}
\title{Calculate diversity indices for barcode samples}
\usage{
calcDivIndexes(dgeObject, group = NULL)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{group}{Optional, column name in sample metadata to group samples by (string).}
}
\value{
Returns a data-frame containing the calculated diversity indices per sample
}
\description{
Takes a DGEList object of barcode counts and computes shannon,
simpson, inverse simpson and gini coefficients for each
sample
}
\details{
calcDivIndexes

Calculates common diversity indices per sample in a dataframe of counts
}
\examples{
data(test.dge)
indexes <- calcDivIndexes(test.dge)
}
