% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPercentileBarcodes.R
\name{calcPercentileBarcodes}
\alias{calcPercentileBarcodes}
\title{calcPercentileBarcodes}
\usage{
calcPercentileBarcodes(dgeObject, percentile = 0.95)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{percentile}{Desired percentile value (decimal). Default = `0.95` (i.e. 95%).}
}
\value{
Returns a list object containing Nth percentile tables of barcode counts per sample
}
\description{
Calculate barcodes comprising the Nth percentile for each sample & generate cumulative sum plots
}
\examples{
data(test.dge)
calcPercentileBarcodes(test.dge, percentile = 0.95)
}
