% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcReplicateCorr.R
\name{calcReplicateCorr}
\alias{calcReplicateCorr}
\title{Calculate correlation between technical replicates}
\usage{
calcReplicateCorr(
  dgeObject,
  group = NULL,
  corrThreshold = 0.8,
  return = "all",
  method = "pearson",
  ignoreZero = FALSE
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts with technical replicates.}

\item{group}{Column name in sample metadata with replicate information (string). Correlation can only be calculated between pairs of replicates.}

\item{corrThreshold}{Threshold distinguishing good vs bad correlation between technical replicates (decimal). Default = `0.8`.}

\item{return}{Which values to return, one of `good`, `bad`, `all` (string). Default = `all`.}

\item{method}{Method for correlation, one of `spearman`, `pearson` (string). Default = `pearson`.}

\item{ignoreZero}{Remove barcodes where both replicates have 0 counts before calculating correlation (boolean). Including zeros can increase the spearman correlation, especially for samples with few barcodes detected. Default = `FALSE`.}
}
\value{
Returns a plot of the read counts per barcode (row) in a data frame
}
\description{
Calculate and return list of correlation between paired technical replicates in a dataset
}
\examples{
data(test.dge)
calcReplicateCorr(test.dge, group = "group")
}
