% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_cormat}
\alias{cluster_cormat}
\title{Cluster correlation matrix}
\usage{
cluster_cormat(correlation_mat)
}
\arguments{
\item{cormat}{matrix of correlation values}
}
\value{
Returns a matrix of correlation values with columns and rows hierarchically clustered
}
\description{
cluster a correlation matrix using hierarchical clustering
}
