% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseReplicates.R
\name{collapseReplicates}
\alias{collapseReplicates}
\title{Collapse technical replicates}
\usage{
collapseReplicates(
  dgeObject,
  group,
  sampleNames = NULL,
  renameCols = TRUE,
  showReps = FALSE,
  method = "mean"
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{group}{Column name in sample metadata to group samples by (string).}

\item{sampleNames}{Optional, column name in sample metadata that contains unique sample names (string). If provided, a new column collapsed_samples will be added to the sample metadata with the collapsed samples}

\item{renameCols}{Whether to rename the columns of the returned object using the levels of the grouping factor (boolean). Default = `TRUE`.}

\item{showReps}{Whether to print replicate column names to console (boolean). Default = `FALSE`.}

\item{method}{Method to collapse replicates by, one of `mean` or `sum` (string). Default = `mean`.}
}
\value{
Returns a DGE object with normalised counts for a sample.
}
\description{
Collapse technical replicates in a DGEList object by mean or sum.
Modified from collapseReplicates function from DESeq2 package to accept DGEList objects and to allow collapsing replicated by mean or sum.
}
\examples{
data(test.dge)
collapseReplicates(test.dge, group = "group", method = "mean")

}
