% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBarcodeReads.R
\name{countBarcodeReads}
\alias{countBarcodeReads}
\title{Count Barcode Reads}
\usage{
countBarcodeReads(bam, bam_index, outdir = ".", prefix = "map_barcode_reads")
}
\arguments{
\item{bam}{BAM file from mapBarcodeReads containing mapped barcode reads}

\item{bam_index}{Path to bowtie index of barcode reference library. Include index prefix.}

\item{outdir}{Optional. Save output to directory}

\item{prefix}{Optional. Prefix to add to output file}
}
\value{
Returns a data-frame containing the Barcode ID, and the count per sample.
}
\description{
Count number of occurences of each barcode in the reference library
}
