% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dose}
\alias{dose}
\title{dose escalation dataset DGEList object}
\format{
A DGElist object data frame with 13338 rows and 90 variables:
}
\usage{
dose
}
\description{
A dataset containing raw count information for the dose dataset
}
\keyword{datasets}
