% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcodeReads.R
\name{extractBAM}
\alias{extractBAM}
\title{extractBAM}
\usage{
extractBAM(
  bam,
  barcode_pattern,
  constant = "",
  yieldSize = 1e+06,
  outfile = NULL
)
}
\arguments{
\item{bam}{BAM file containing barcoded reads}

\item{barcode_pattern}{Pattern to search for in the reads}

\item{constant}{The constant region flanking the barcode, two constant regions can be input as a
character vector. This should be set if the constant region is part of the barcode_pattern.
Up to two constant regions can be set e.g. c("ACTG","TACG")}

\item{yieldSize}{How many lines of input file to stream at each pass}

\item{outfile}{Optional. path to save parsed reads as output file}
}
\value{
Returns a data-frame containing the read ID, the 10X cell ID, 10X umi, lintraceR DNA
barcode sequence and lintraceR DNA barcode ID.
}
\description{
Barcode extraction function for BAM alignment files
}
