% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcodeReads.R
\name{extractBarcodeReads}
\alias{extractBarcodeReads}
\title{extractBarcodeReads}
\usage{
extractBarcodeReads(
  infile,
  barcode_pattern = "[ACTG][ACTG][GC][AT][GC][ACTG][ACTG][AT][GC][AT]",
  constant = NULL,
  yieldSize = 1e+06,
  outfile = NULL,
  clean = TRUE,
  verbose = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{infile}{Input fastq or bam file containing barcoded reads}

\item{barcode_pattern}{Pattern to search for in the reads}

\item{constant}{The constant region flanking the barcode, two constant regions can be input as a
character vector. This should be set if the constant region is part of the barcode_pattern.
Up to two constant regions can be set e.g. c("ACTG","TACG")}

\item{yieldSize}{How many lines of input file to stream at each pass}

\item{outfile}{Optional. path to save parsed reads as output file}

\item{clean}{Logical. remove previous output files generated by extractBarcodeReads before current run}

\item{verbose}{Logical. print detailed output to console}

\item{overwrite}{Logical. Overwrite previous output files generated by extractBarcodeReads}
}
\value{
Returns a data-frame containing the read ID, the 10X cell ID, 10X umi, lintraceR DNA
barcode sequence and lintraceR DNA barcode ID.
}
\description{
Identify and extract reads containing DNA barcodes from a FASTQ or BAM alignment file
}
\examples{
test.fq <- system.file("extdata", "test_extract.fastq.gz", package = "barista")
extractBarcodeReads(infile = test.fq, barcode_pattern = "ACTG")
test.bam <- system.file("extdata", "test_extract_anno.bam", package = "barista")
extractBarcodeReads(infile = test.bam, barcode_pattern = "ACTG")
}
