% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcodeReads.R
\name{extractFASTQ}
\alias{extractFASTQ}
\title{extractFASTQ}
\usage{
extractFASTQ(
  fastq,
  barcode_pattern,
  constant = NULL,
  outfile = NULL,
  block_size = 1e+06,
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{fastq}{REQUIRED. Path to input fastq file}

\item{barcode_pattern}{REQUIRED. Pattern to search for in the reads.}

\item{constant}{The constant region(s) flanking the barcode. This should be set if the
constant region is part of the barcode_pattern.
Up to two constant regions can be set e.g. constant = c("ACTG","TACG")}

\item{outfile}{OPTIONAL. path to save filtered reads as output fasta file with read name as sequence ID}

\item{block_size}{Number of reads to load into memory per chunk. Default = 5e5}

\item{verbose}{Boolean. Show verbose output from ShortRead? Default = FALSE.}

\item{overwrite}{Boolean. Overwrite output files? Default = FALSE.}
}
\value{
Returns a data-frame containing the read ID and the extracted DNA
barcode sequence
}
\description{
Barcode extraction function for raw FASTQ files
}
