% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBARtabQC.R
\name{filterBarcodes}
\alias{filterBarcodes}
\title{filterBarcodes}
\usage{
filterBarcodes(counts, umi_count_filter = 1, umi_fraction_filter = 0.3)
}
\arguments{
\item{counts}{Dataframe with barcodes and UMI counts per cell}

\item{umi_count_filter}{Minimum number of UMIs per barcode per cell}

\item{umi_fraction_filter}{Minimum fraction of UMIs per barcode per cell compared to dominant barcode in cell (barcode supported by most UMIs)}
}
\value{
Returns a data frame with one row per cell ID
}
\description{
Filter barcodes from single-cell sample
}
