% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeDist.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{Get upper triangle}
\usage{
get_upper_tri(cormat)
}
\arguments{
\item{cormat}{matrix of correlation values}
}
\value{
Returns a matrix of correlation values with lower triangle values changed to NA
}
\description{
change lower triangle values in a mirrored matrix to NA
}
