% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominantBarcodes.R
\name{getDominantBarcodes}
\alias{getDominantBarcodes}
\title{Get dominant barcodes per sample}
\usage{
getDominantBarcodes(dgeObject, threshold = 0.05)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{threshold}{Proportion threshold to call dominant barcodes (decimal). Default = `0.05` (i.e. 5%).}
}
}
\value{
Returns a named list containing vectors of dominant barcodes per sample
}
\description{
Takes a DGEList or dataframe of barcode counts,
computes the proportion abundance of each barcode within each sample and
then returns a list of barcodes meeting a threshold abundance per sample.
}
\examples{
data(test.dge)
getDominantBarcodes(test.dge, threshold = 0.05)
}
