% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBarcodeReads.R
\name{mapBarcodeReads}
\alias{mapBarcodeReads}
\title{Map Barcode Reads}
\usage{
mapBarcodeReads(
  reads,
  bowtie_index = NULL,
  mismatches = 2,
  threads = 4,
  outdir = getwd(),
  prefix = "map_barcode_reads",
  cleanup = TRUE,
  reference_fasta = NULL
)
}
\arguments{
\item{reads}{Dataframe from extractBarcodeReads containing barcode reads}

\item{bowtie_index}{Path to bowtie index of barcode reference library. Include index prefix.}

\item{mismatches}{Number of mismatches allowed in bowtie for read mapping}

\item{threads}{Number of CPU threads for bowtie to use}

\item{outdir}{Optional. save output to directory}

\item{prefix}{Optional. prefix to add to output file}

\item{cleanup}{Logical. Remove intermediate files generated by mapBarcodeReads}

\item{reference_fasta}{Optional. Fasta file of reference barcode library for alignment. Use instead of bowtie_index.}
}
\value{
Returns a data-frame containing the read ID, the 10X cell ID, 10X umi, lintraceR DNA
barcode sequence mapped to the lintraceR DNA barcode ID.
}
\description{
Map reads containing a lintraceR DNA barcode to a barcode reference library/fasta and return CellID:lintracer barcode ID annotations
}
\examples{
load(system.file("extdata", "test_map.rda", package = "barista"))
test.ref <- system.file("extdata", "barcode_lib_reference_test.fasta", package = "barista")
mapBarcodeReads(reads = test.map, reference_fasta = test.ref, outdir = tempdir(), prefix = "map")
}
