% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseCounts.R
\name{normaliseCounts}
\alias{normaliseCounts}
\title{Normalise barcode counts}
\usage{
normaliseCounts(dge, method = "CPM", threshold = 0)
}
\arguments{
\item{dge}{DGEList object containing raw barcode counts to be normalised}

\item{method}{method of normalization. One of cpm, tmm, tmmwsp, upperquartile or rle. see edgeR::calcNormFactors()}

\item{threshold}{Threshold to apply to counts before normalisation}
}
\value{
Returns a data.frame object containing normalised barcode counts
}
\description{
Normalise barcode counts within a DGEList object
}
\examples{
data(test.dge)
normaliseCounts(test.dge, method = "CPM", threshold = 0)

}
