% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBoxplot.R
\name{plotBarcodeBoxplot}
\alias{plotBarcodeBoxplot}
\title{Boxplot of selected barcodes}
\usage{
plotBarcodeBoxplot(
  dge.obj,
  barcodes = NULL,
  group = NULL,
  conditions = NULL,
  trans = NULL,
  point = FALSE,
  violin = FALSE
)
}
\arguments{
\item{dge.obj}{DGE object containing counts and sample information}

\item{barcodes}{a vector of barcodes to be plotted}

\item{group}{metadata field to facet data on}

\item{conditions}{specific levels of group to be plotted. If NULL all levels are plotted}

\item{trans}{From ggplot2. For continuous scales, the name of a transformation object or the object itself.}

\item{point}{From ggplot2. Include plot points?}

\item{violin}{From ggplot2. Include violin plot?}
}
\value{
Returns a boxplot
}
\description{
Takes a dataframe of barcode counts,
select specific barcodes in the dataframe
and boxplot their counts
}
\details{
plotBCBoxplot

Boxplot of selected barcodes
}
\examples{
data(test.dge)
plotBarcodeBoxplot(test.dge, barcodes = "BC_190202", group = "Treatment", 
conditions = c("Vehicle", "Low_dose", "High_dose"))
}
