% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeCorrelation.R, R/plotBarcodeDist.R
\name{plotBarcodeCorrelation}
\alias{plotBarcodeCorrelation}
\title{Sample to sample correlation}
\usage{
plotBarcodeCorrelation(
  counts,
  method = "pearson",
  upper = T,
  clustered = T,
  name = "Sample Correlation"
)

plotBarcodeCorrelation(
  counts,
  method = "pearson",
  upper = T,
  clustered = T,
  name = "Sample Correlation"
)
}
\arguments{
\item{counts}{matrix/dataframe containing raw or normalised counts}

\item{method}{correlation metric to use. Refer to stats::cor for available options}

\item{upper}{Logical. plot only the upper half of the matrix}

\item{clustered}{Logical. cluster rows and columns of matrix}

\item{name}{title of plot}

\item{dgeObject}{DGEList object with barcode counts.}

\item{title}{Title of plot (string). Default = `Sample Correlation - method`.}
}
\value{
Returns a heatmap of sample distances using desired clustering

Returns a heatmap of sample distances using desired clustering
}
\description{
Plot sample correlation between barcode sets / samples

Plot sample correlation between barcode sets / samples
}
\examples{
data(test.dge)
plotBarcodeCorrelation(test.dge)
test.mat <- matrix(rnorm(10*10,mean=0,sd=2), 10, 10)
plotBarcodeCorrelation(counts = test.mat, method = "pearson")
}
