% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeCounts.R
\name{plotBarcodeCounts}
\alias{plotBarcodeCounts}
\title{Plot total counts per barcode in library}
\usage{
plotBarcodeCounts(dgeObject, order = FALSE, log10 = FALSE)
}
\arguments{
\item{dgeObject}{DGE object with barcode counts}

\item{order}{Order the dataset by decreasing abundance (boolean). Default = `FALSE`.}

\item{log10}{log10 transform the data. Adds a pseudocount of 1 (boolean). Default = `FALSE`.}
}
\value{
Returns a plot of the read counts per barcode (row) in a data frame
}
\description{
Simple plot of total read counts per barcode across samples.
}
\examples{
data(test.dge)
plotBarcodeCounts(test.dge)
plotBarcodeCounts(test.dge, order = TRUE, log10 = TRUE)
}
