% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeDist.R, R/plotBarcodeDistance.R
\name{plotBarcodeDistance}
\alias{plotBarcodeDistance}
\title{Sample to sample distances}
\usage{
plotBarcodeDistance(
  dgeObject,
  method = "euclidean",
  upper = TRUE,
  clustered = TRUE,
  title = "Sample Distances"
)

plotBarcodeDistance(
  dgeObject,
  method = "euclidean",
  upper = TRUE,
  clustered = TRUE,
  title = "Sample Distances"
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{method}{Distance metric to use (string). One of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Refer to stats::dist for details. Default = `euclidean`.}

\item{upper}{Plot only the upper half of the matrix (boolean). Default = `TRUE`.}

\item{clustered}{Cluster rows and columns of matrix (boolean). Default = `TRUE`.}

\item{title}{Title of plot (string). Default = `Sample Distances - method`.}

\item{counts}{matrix/dataframe containing raw or normalised counts}

\item{name}{title of plot}
}
\value{
Returns a heatmap of sample distances using desired clustering

Returns a heatmap of sample distances using desired clustering
}
\description{
Plot sample distances between barcode sets / samples

Plot sample distances between barcode sets / samples
}
\examples{
plotBarcodeDistance(test.dge$counts)
data(test.dge)
plotBarcodeDistance(test.dge)
}
