% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeHeatmap.R
\name{plotBarcodeHeatmap}
\alias{plotBarcodeHeatmap}
\title{Plot barcode heatmap}
\usage{
plotBarcodeHeatmap(
  counts,
  N,
  name = "CPM",
  show_bc = FALSE,
  samples = NULL,
  group = NULL,
  col_annot = NULL,
  discrete = F,
  discrete_threshold = 1
)
}
\arguments{
\item{counts}{Dataframe containing cpm counts}

\item{N}{Number of top barcodes per sample}

\item{name}{String of the heatmap scale name}

\item{show_bc}{Boolean to show barcode names on the heatmap}

\item{samples}{Dataframe containing sample metadata sheet}

\item{group}{metadata fields to annotate samples}

\item{col_annot}{list of vectors assigning colours to each level of metadata}

\item{discrete}{Boolean to show presence/absence of barcodes instead of abundance}

\item{discrete_threshold}{Threshold for presence of barcode in samples}
}
\value{
Returns a heatmap
}
\description{
Takes a dataframe of the barcode counts, and selects the N most abundant barcodes per sample.
Most abundant barcodes are gathered in a list and a heatmap for those barcodes is plotted for all samples.
Stars indicate the most abundant barcodes per sample.
}
\details{
plotBarcodeHeatmap

Plot a heatmap of the N most abundant barcodes per sample
}
\examples{
col_meta = list(Experiment = c("test_01" = "black"), 
Treatment = c("T0" = "green", "Vehicle" = "purple",
"Low_dose" = "yellow","High_dose" = "brown"))
plotBarcodeHeatmap(counts = cpm(test.dge$counts),N = 5,show_bc = TRUE,
samples = test.dge$samples,group = c("Treatment","Experiment"), 
col_annot = col_meta, discrete = TRUE, discrete_threshold = 1)
}
