% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeHistogram.R
\name{plotBarcodeHistogram}
\alias{plotBarcodeHistogram}
\title{plotBarcodeHistogram}
\usage{
plotBarcodeHistogram(
  dgeObject,
  topN = 10,
  seedColors = 1,
  samples = NULL,
  orderSamples = NULL,
  alphaLowFreq = 1
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{topN}{Number of most frequent barcodes to show in legend (integer). Default = `10`.}

\item{seedColors}{Seed for sampling colors (integer). Default = `1`.}

\item{samples}{Samples to plot (vector of strings). Default all.}

\item{orderSamples}{One or multiple samples to order barcodes by (string or vector of strings). Default all.}

\item{alphaLowFreq}{Alpha of barcodes not in top n barcodes (decimal). Default = `1`.}
}
\value{
Returns a stacked bar plot of barcode frequencies within samples
}
\description{
Generate proportional stacked bar plot of barcodes from raw count object with n most frequent barcodes labelled.
}
\examples{
plotBarcodesStackedBar(test.dge)
}
