% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodePCA.R
\name{plotBarcodePCA}
\alias{plotBarcodePCA}
\title{Barcode PCA plot}
\usage{
plotBarcodePCA(
  dgeObject,
  groups = NULL,
  ntop = 500,
  returnData = FALSE,
  pcs = c(1, 2),
  batch = NULL
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{groups}{Optional, one or multiple column names in sample metadata to color samples by (string or vector of strings).}

\item{ntop}{number of top most variable barcodes to be used in PCA calculation (integer). Default = `500`.}

\item{returnData}{Return a data.frame of PCA calculation instead of plot (boolean). Default = `FALSE`.}

\item{pcs}{Principle components to plot (vector of integers of length 2). Default = `c(1, 2)`.}

\item{batch}{Optional, metadata category indicating source of batch effects. Will be used in limma removeBatchEffect function prior to PCA.}
}
\value{
Returns a plot of the first two principal components in the dataset
}
\description{
Takes a DGEList object contining barcode cpm counts and performs PCA.
Then, plots the first two dimensions.
Modified from DESeq2 package plotPCA function
}
\details{
plotBarcodePCA

Plot the first two principal components of the barcode dataset
}
\examples{
data(test.dge)
plotBarcodePCA(test.dge, groups = "Treatment", ntop = 500, returnData = FALSE, batch = NULL)
}
