% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeRegression.R
\name{plotBarcodeRegression}
\alias{plotBarcodeRegression}
\title{plotBarcodeRegression}
\usage{
plotBarcodeRegression(
  dgeObject,
  sample1 = NULL,
  sample2 = NULL,
  title = NULL,
  trendline = TRUE,
  trans = NULL,
  rug = FALSE
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{sample1}{Name of sample 1 (string).}

\item{sample2}{Name of sample 2 (string).}

\item{title}{Optional, title of plot (string).}

\item{trendline}{Include linear trendline using `stat_smooth()` (boolean). Default = `TRUE`.}

\item{trans}{Optional, the name of a transformation object or the object itself.}

\item{rug}{Include geom_rug density information on the axes (boolean). Defaule = `FALSE`.}
}
\value{
Returns a scatterplot of the two samples
}
\description{
Generate a linear regression scatterplot for two sets of sample counts.
}
\examples{
data(test.dge)
plotBarcodeRegression(test.dge, sample1 = "T0-1", sample2 = "T0-2")
}
