% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBARtabQC.R
\name{plotBarcodesPerCell}
\alias{plotBarcodesPerCell}
\title{plotBarcodesPerCell}
\usage{
plotBarcodesPerCell(counts, fraction = TRUE)
}
\arguments{
\item{counts}{Dataframe with barcodes and UMI counts per cell}

\item{fraction}{Boolean, whether to plot fraction or number of cells}
}
\value{
Returns a plot
}
\description{
Plot number of detected barcodes per cell
}
