% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeTimeseries.R
\name{plotBarcodeTimeseries}
\alias{plotBarcodeTimeseries}
\title{plotBarcodeTimeseries}
\usage{
plotBarcodeTimeseries(counts, name = "", seed = 5, top = 50)
}
\arguments{
\item{counts}{dataframe containing raw counts of barcodes}

\item{name}{desired plot title}

\item{seed}{RNG seed}

\item{top}{number of top barcodes per sample to plot}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate proportional timeseries plot from raw / normalised barcode count object.
}
\examples{
data(test.dge)
plotBarcodeTimeseries(test.dge[,1:4], name = "Proportional Timeseries Plot", seed = 5)
}
