% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBARtabQC.R
\name{plotBARtabFilterQC}
\alias{plotBARtabFilterQC}
\title{plotBARtabFilterQC}
\usage{
plotBARtabFilterQC(
  dir = NULL,
  recursive = T,
  patternLog = "*filter.log",
  fullNames = T,
  normalised = F,
  plot = T,
  title = "BARtab Filter QC"
)
}
\arguments{
\item{dir}{directory where BARtab was successfuly run on barcode count datasets}

\item{recursive}{Logical. TRUE will recurse regex search into subdirectories}

\item{patternLog}{regex string to specify filter stage log files}

\item{fullNames}{Logical. Return full names of files detected by regex search}

\item{normalised}{Logical. log10 normalise counts}

\item{plot}{Logical. Generate plots. False returns raw data}

\item{title}{Optional. title of plots.}
}
\value{
Returns a plot / dataset containing BARtab Filtering stage QC data
}
\description{
Plot filtered read percentages from a BARtab run folder
}
