% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsinClusters.R
\name{plotCellsInClusters}
\alias{plotCellsInClusters}
\title{Plot distribution of group across clusters}
\usage{
plotCellsInClusters(
  sc.obj,
  group = "barcode",
  factor = NULL,
  clusters = "seurat_clusters",
  plot.pct = T,
  plot = T
)
}
\arguments{
\item{sc.obj}{Single cell Seurat or SingleCellExperiment object containing clusters and group metadata.}

\item{group}{a column of metadata (string). Default = `barcode`.}

\item{factor}{a level of group to test for enrichment per cluster, e.g. a specific barcode (string).}

\item{clusters}{a column of metadata defining the cluster identities of each cell (string). Default = `seurat_clusters`}

\item{plot.pct}{Plot percentages (`TRUE`) or raw numbers (`FALSE`) (boolean). Default = `TRUE`}

\item{plot}{Create plot (`TRUE`) or return data (`FALSE`) (boolean). Default = `TRUE`.}
}
\value{
Returns a histogram or underlying plot data
}
\description{
Takes a single cell object, a grouping variable, a factor within the group to test,
and an ident class (i.e. clusters). Plots percentage or raw number of cells within factor
present per level of ident (i.e. per cluster)
}
\details{
Plot number or percentage of a group per level of ident in a single
cell object
}
