% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsPerGroup.R
\name{plotCellsPerGroup}
\alias{plotCellsPerGroup}
\title{Plot of cells per barcode}
\usage{
plotCellsPerGroup(
  sc.obj = NULL,
  group = NULL,
  order = TRUE,
  trans = NULL,
  threshold = 100,
  plot = TRUE,
  label = TRUE,
  sep = ";"
)
}
\arguments{
\item{sc.obj}{single cell object in Seurat or SingleCellExperiment format containing group metadata}

\item{group}{a column of metadata in sc.obj}

\item{order}{Logical. Rank order levels of group by cell number}

\item{trans}{From ggplot2. For continuous scales, the name of a transformation object or the object itself.}

\item{threshold}{threshold number of cells above which labels for group levels will appear}

\item{plot}{Logical. Plot results or return data.}

\item{label}{Logical. Label group levels above cell number threshold. if ggrepel is installed will use geom_text_repel
instead of geom_text}

\item{sep}{Separating character used for aggregation (string). Default = `;`.}
}
\value{
Returns a plot of cell number by barcode test results or underlying plot data
}
\description{
Takes a single cell object, a grouping variable, a factor within the group to test,
and an ident class (i.e. clusters). Per level of ident performs hypergeometric testing
for enrichment of group factor
}
\details{
plotCellsPerGroup

Plots number of cells for each detected barcode in a single cell dataset
}
