% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDetectedBarcodes.R
\name{plotDetectedBarcodes}
\alias{plotDetectedBarcodes}
\title{plotDetectedBarcodes}
\usage{
plotDetectedBarcodes(
  dgeObject,
  percentile = 0.95,
  plot = T,
  sampleOrder = NULL,
  group = NULL,
  title = NULL
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{percentile}{Percentile threshold to count barcodes (decimal). Default = `0.95`.}

\item{plot}{Plot data instead of returning counts table (boolean). Default = `TRUE`.}

\item{sampleOrder}{Optional, Ordering of samples (vector of strings).}

\item{group}{Optional, column name in sample metadata to group samples by (string).}

\item{title}{Optional, Plot title (string).}
}
\value{
Returns a histogram plot of the number of detected barcodes per sample.
}
\description{
Plot the total number of barcodes detected in a sample above a percentile threshold.
}
\examples{
data(test.dge)
plotDetectedBarcodes(test.dge, percentile = .95)
plotDetectedBarcodes(test.dge, plot = FALSE)
}
