% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivIndexes.R
\name{plotDivIndexes}
\alias{plotDivIndexes}
\title{Plot diversity indices across barcode samples}
\usage{
plotDivIndexes(
  dgeObject,
  div = NULL,
  metric = "shannon",
  type = "bar",
  group = NULL
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{div}{Optional, precomputed diversity metrics calculated by `calcDivIndexes()`.}

\item{metric}{Diversity metric to plot (string). One of "shannon", "simpson", "invsimpson" or "gini". Default = `shannon`.}

\item{type}{Plot as bar, point or box plot (string). Default = `bar`.}

\item{group}{Optional, column name in sample metadata to group samples by (string).}
}
\value{
Returns a plot of calculated diversity index per sample
}
\description{
Plots a dataframe of diversity indices per sample
}
\details{
plotDivIndexes

Plots common diversity indices per sample
}
\examples{
data(test.dge)
plotDivIndexes(test.dge, group="Treatment", type="box")
}
