% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLibraryMetrics.R
\name{plotLibraryCumSum}
\alias{plotLibraryCumSum}
\title{Cumulative sum plot of barcode abundances in a reference library}
\usage{
plotLibraryCumSum(barcodes, samplename = "Library", cutoff = 10)
}
\arguments{
\item{barcodes}{dataframe containing raw counts of barcodes}

\item{samplename}{sample condition of interest}

\item{cutoff}{rowsum cutoff defining rows to keep}
}
\value{
Returns a cumulative sum plot for the desired barcode library sample
}
\description{
Takes a dataframe of barcode counts,
computes the median abundance of each barcode for two specific conditions,
then do a line plot for both conditions
}
\details{
plotLibraryCumSum

Plot the cumulative sum of the barcode abundances in a barcode library
}
