% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLibraryMetrics.R
\name{plotLibraryDiversity}
\alias{plotLibraryDiversity}
\title{Lineplot of barcode abundances in two conditions}
\usage{
plotLibraryDiversity(barcodes, samplename = "Library", cutoff = 10, skew = F)
}
\arguments{
\item{barcodes}{dataframe containing raw counts of barcodes}

\item{samplename}{sample condition of interest}

\item{cutoff}{rowsum cutoff defining rows to keep}

\item{skew}{Logical. Calculate and plot the library skew ratio (90th percentile / 10th percentile)}
}
\value{
Returns a barcode library frequency plot
}
\description{
Takes a dataframe of barcode counts,
computes and plots the abundance of each barcode in the library sample
}
\details{
plotLibraryDiversity

Lineplot of the barcode abundances in two different experimental settings
}
