% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{plotMetrics}
\alias{plotMetrics}
\title{Violin plot of selected metric per group variable}
\usage{
plotMetrics(
  sc.obj = NULL,
  group = NULL,
  trans = NULL,
  factor = NULL,
  threshold = 100,
  plot = TRUE
)
}
\arguments{
\item{sc.obj}{single cell object in Seurat or SingleCellExperiment format containing group metadata}

\item{group}{a column of metadata in sc.obj containing grouping information}

\item{trans}{From ggplot2. For continuous scales, the name of a transformation object or the object itself.}

\item{factor}{a metadata field to plot per level of group}

\item{threshold}{threshold number of cells per level of group to plot}

\item{plot}{Logical. Plot results or return data.}
}
\value{
Returns a plot of cell number by barcode test results or underlying plot data
}
\description{
Takes a single cell object, a grouping variable and a metadata variable.
Plots distribution of factor per level of group with total number of cells above threshold
}
\details{
plotMetrics

Plots metric distribution for selected groups in a single cell dataset
}
