% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrderedBubble.R
\name{plotOrderedBubble}
\alias{plotOrderedBubble}
\title{plotOrderedBubble}
\usage{
plotOrderedBubble(
  counts,
  title = "Proportional Bubble Plot",
  orderSample = NULL,
  samples = NULL,
  group = NULL,
  displaySamples = NULL,
  displayBarcodes = NULL,
  proportionCutoff = 10,
  colorDominant = FALSE,
  filterCutoff = NULL,
  labelBarcodes = TRUE,
  legend = TRUE
)
}
\arguments{
\item{counts}{dataframe containing raw counts of barcodes. Expects barcodes as row names and samples as columns. Alternatively, DGE object.}

\item{title}{desired plot title}

\item{orderSample}{name of sample to order by}

\item{samples}{Dataframe containing sample metadata sheet with samples as row names.}

\item{group}{metadata field to annotate samples on bubble plot}

\item{displaySamples}{vector of samples to display - keep the order of vector}

\item{displayBarcodes}{vector of barcodes to display}

\item{proportionCutoff}{barcodes represented at a percentage within any sample above this threshold will be labelled}

\item{colorDominant}{only color clones with frequency above proportionCutoff. Others colored grey}

\item{filterCutoff}{barcodes below this threshold in orderSample will be filtered in all samples}

\item{labelBarcodes}{Boolean, whether to label barcodes with a proportion larger than proportionCutoff in any sample}

\item{legend}{Boolean, whether to show a legend of bubble sizes}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate ordered proportional bubbleplots from raw count object with barcodes labelled above a specified threshold
}
\examples{
plotOrderedBubble(test.dge$counts, orderSample = "T0-1", filterCutoff = 0.001)
}
