% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadCounts.R
\name{plotReadCounts}
\alias{plotReadCounts}
\title{Plot sample read counts}
\usage{
plotReadCounts(
  dgeObject,
  group = NULL,
  log10 = FALSE,
  legend = TRUE,
  order = TRUE
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{group}{Optional, column name in sample metadata to group samples by (string).}

\item{log10}{log10 transform output (boolean). Default = `FALSE`.}

\item{legend}{Include legend (boolean). Default = `TRUE`.}

\item{order}{Order samples by group (boolean). Default = `TRUE`.}
}
\value{
Returns a plot of the read counts per column (sample) in a data frame
}
\description{
Simple plot of total read counts per sample
}
\examples{
data(test.dge)
plotReadCounts(test.dge, group = "Treatment")

}
