% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleCumSum.R
\name{plotSampleCumSum}
\alias{plotSampleCumSum}
\title{Plot the cumulative sum of barcode abundances within samples to visualize sample diversity.}
\usage{
plotSampleCumSum(dgeObject, samples = NULL)
}
\arguments{
\item{dgeObject}{DGE object with barcode counts}

\item{samples}{Sample names}
}
\value{
Returns a cumulative sum plot
}
\description{
plotSampleCumSum
}
