% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellBARtabQC.R
\name{plotUmiFilterThresholds}
\alias{plotUmiFilterThresholds}
\title{plotUmiFilterThresholds}
\usage{
plotUmiFilterThresholds(counts)
}
\arguments{
\item{counts}{Dataframe with barcodes and UMI counts per cell}
}
\value{
Returns a plot
}
\description{
Plot number of cells with barcodes annotated given different UMI count filter thresholds
}
