% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBarcodeLibrary.R
\name{processBarcodeLibrary}
\alias{processBarcodeLibrary}
\title{processBarcodeLibrary}
\usage{
processBarcodeLibrary(
  file = NULL,
  samplename = "Barcode",
  cutoff = 10,
  header = FALSE,
  outdir = tempdir()
)
}
\arguments{
\item{file}{path to starcode output for barcode library}

\item{samplename}{name of reference library. Will be prefixed to each barcode in the reference}

\item{cutoff}{rowsum cutoff defining barcodes to keep in the reference library}

\item{header}{Logical. Does the input file contain a header line}

\item{outdir}{Desired output directory to save library files}
}
\value{
returns data frame containing barcode and read count in reference
}
\description{
process barcode reference file from raw sequencing datasets
}
\examples{
load(system.file("extdata", "test_raw_lib.rda", package = "barista"))
processBarcodeLibrary(file = test.raw.lib, samplename = "Barcode", outdir = tempdir())
}
