% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionalBubbleplot.R
\name{proportionalBubbleplot}
\alias{proportionalBubbleplot}
\title{proportionalBubbleplot}
\usage{
proportionalBubbleplot(
  counts.obj,
  labels = TRUE,
  proportion.cutoff = 10,
  name = "Proportional Bubble Plot"
)
}
\arguments{
\item{counts.obj}{dataframe containing raw counts of barcodes. assumes barcodes as rownames.}

\item{labels}{Boolean. print barcode labels?}

\item{proportion.cutoff}{barcodes represented at a percentage within any sample above this threshold will be labelled}

\item{name}{desired plot title}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate proportional bubbleplots from raw count object with barcodes labelled above a specified threshold
}
\examples{
data(test.counts)
proportionalBubbleplot(test.counts, name = "Proportional Bubble Plot", proportion.cutoff = 10)
}
