% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBARtabQC.R
\name{readBartabCounts}
\alias{readBartabCounts}
\title{readBartabCounts}
\usage{
readBartabCounts(counts_path)
}
\arguments{
\item{counts_path}{Path to BARtab counts.tsv file}
}
\value{
Returns a data frame in long format of cell IDs, barcodes and barcode UMI counts
}
\description{
Read BARtab counts table from single-cell sample
}
