% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.bartab.sc}
\alias{test.bartab.sc}
\title{Test BARtab single cell barcode annotation}
\format{
A data frame with 3000 rows and 3 variables:
}
\usage{
test.bartab.sc
}
\description{
A table of cell id, barcode, UMI counts
}
\keyword{datasets}
