% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.sce}
\alias{test.sce}
\title{Test barcode SCE object}
\format{
A SCE object with 100 rows and 100 variables:
}
\usage{
test.sce
}
\description{
A SingleCellExperiment object containing single cell RNA-seq data for 100 cells
and 100 features
}
\keyword{datasets}
