% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDivIndexes.R, R/plotDivIndexes.R
\name{calcDivIndexes}
\alias{calcDivIndexes}
\title{Calculate diversity indices for barcode samples}
\usage{
calcDivIndexes(counts, div = NULL)

calcDivIndexes(counts, div = NULL)
}
\arguments{
\item{counts}{dataframe of raw or normalised counts with samples as columns and observations/barcodes as rows}

\item{div}{dataframe of diversity metrics as columns and samples as rows}

\item{group}{vector of sample group metadata, assumes the same sample order as counts}
}
\value{
Returns a data-frame containing the calculated diversity indices per sample

Returns a plot of calculated diversity indices per sample
}
\description{
Takes a dataframe of barcode counts and computes shannon,
simpson, inverse simpson and gini coefficients for each
sample

Plots a dataframe of diversity indices per sample
}
\details{
calcDivIndexes

Calculates common diversity indices per sample in a dataframe of counts

plotDivIndexes

Plots common diversity indices per sample
}
\examples{
indexes <- calcDivIndexes(counts = test.counts)
plotDivIndexes(dat = test.counts)
}
