% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDivIndexes.R
\name{calcDivIndexes}
\alias{calcDivIndexes}
\title{Calculate diversity indices for barcode samples}
\usage{
calcDivIndexes(dat)
}
\arguments{
\item{dat}{dataframe of raw or normalised counts with samples as columns and observations/barcodes as rows}
}
\value{
Returns a data-frame containing the calculated diversity indices per sample
}
\description{
Takes a dataframe of barcode counts and computes shannon,
simpson, inverse simpson and gini coefficients for each
sample
}
\details{
calcDivIndexes

Calculates common diversity indices per sample in a dataframe of counts
}
\examples{
indexes <- calcDivIndexes(dat = test.counts)
}
