% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPercentileBarcodes.R
\name{calcPercentileBarcodes}
\alias{calcPercentileBarcodes}
\title{calcPercentileBarcodes}
\usage{
calcPercentileBarcodes(counts.obj, percentile = 0.95)
}
\arguments{
\item{counts.obj}{DGEList object containing raw or normalised barcode counts}

\item{percentile}{desired percentile value}
}
\value{
Returns a list object containing nth percentile tables of barcode counts per sample
}
\description{
Calculate barcodes comprising the Nth percentile for each sample & generate cumulative sum plots
}
\examples{
data(test.dge)
calcPercentileBarcodes(test.dge, percentile = .95)
}
