% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcReplicateCorr.R
\name{calcReplicateCorr}
\alias{calcReplicateCorr}
\title{Calculate correlation between technical replicates}
\usage{
calcReplicateCorr(dge, group, corr.thresh = 0.8, return = "all")
}
\arguments{
\item{dge}{DGEList object containing technical replicates to be compared}

\item{group}{grouping variable from metadata containing technical replicate information}

\item{corr.thresh}{threshold distinguishing good vs bad correlation between technical replicates}

\item{return}{which values to return. One of "good", "bad", "all"}
}
\value{
Returns a plot of the read counts per barcode (row) in a data frame
}
\description{
Calculate and return list of correlation between paired technical replicates in a dataset
}
\examples{
data(test.dge)
calcReplicateCorr(test.dge, group = "group")
}
