% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseReplicates.R
\name{collapseReplicates}
\alias{collapseReplicates}
\title{Collapse technical replicates}
\usage{
collapseReplicates(
  object,
  groupby,
  run,
  renameCols = TRUE,
  show_reps = TRUE,
  by = "mean"
)
}
\arguments{
\item{object}{DGEList object containing raw or normalised barcode counts with replicate grouping info in object metadata}

\item{groupby}{a character vector of containing grouping information. Must be as long as the columns of object. Can pass a metadata column from object.}

\item{run}{optional, the names of each unique column in object. if provided, a new column runsCollapsed will be added to the colData which pastes together the names of run}

\item{renameCols}{boolean. whether to rename the columns of the returned object using the levels of the grouping factor}

\item{show_reps}{boolean. print replicate column names to console?}

\item{by}{collapse replicates by mean or sum}
}
\value{
Returns a dataframe of normalised counts for a sample
}
\description{
Collapse technical replicates in a DGEList object by mean or sum.
Modified from collapseReplicates function from DESeq2 package to accept DGEList objects and to allow collapsing replicated by mean or sum.
}
\examples{
data(test.dge)
collapseReplicates(test.dge, groupby = test.dge$samples$group, by = "mean")

}
