% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareAbundance.R
\name{compareAbundance}
\alias{compareAbundance}
\title{Barcode abundance comparison}
\usage{
compareAbundance(counts, condition, condition_names)
}
\arguments{
\item{counts}{dataframe containing raw counts of barcodes}

\item{condition}{sample condition of interest}

\item{condition_names}{names of possible conditions in experiment}
}
\value{
Returns a text file
}
\description{
Takes a dataframe of barcode counts,
computes the mean abundance of each barcode for two specific conditions,
then do a line plot for both conditions
}
\details{
compareAbundance

Compare barcode abundance between 2 conditions using EdgeR
}
\examples{
data(test.dge)
compareAbundance(test.dge, test.dge$samples$group, c("T0","T0b"))
}
