% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellBARtabQC.R
\name{filterBarcodes}
\alias{filterBarcodes}
\title{filterBarcodes}
\usage{
filterBarcodes(counts, umiCountFilter = 1, umiFractionFilter = 0.3)
}
\arguments{
\item{counts}{Dataframe with barcodes and UMI counts per cell}

\item{umiCountFilter}{Minimum number of UMIs per barcode per cell}

\item{umiFractionFilter}{Minimum fraction of UMIs per barcode per cell compared to dominant barcode in cell (barcode supported by most UMIs)}
}
\value{
Returns a data frame with one row per cell ID
}
\description{
Filter barcodes from single-cell sample
}
