% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominantBarcodes.R
\name{getDominantBarcodes}
\alias{getDominantBarcodes}
\title{Get dominant barcodes per sample}
\usage{
getDominantBarcodes(counts.obj, pct.thresh = 5)
}
\arguments{
\item{counts.obj}{DGEList object or dataframe containing raw / normalised counts of barcodes}

\item{pct.thresh}{percentage abundance threshold to call dominant barcodes}
}
\value{
Returns a named list containing vectors of dominant barcodes per sample
}
\description{
Takes a DGEList or dataframe of barcode counts,
computes the percentage abundance of each barcode within each sample
then returns a list of barcodes meeting a threshold abundance per sample
}
\details{
getDominantBarcodes

get names of barcodes over a given percentage frequency per sample
}
\examples{
data(test.dge)
getDominantBarcodes(test.dge, pct.thresh = 5)
}
