% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseCounts.R
\name{normaliseCounts}
\alias{normaliseCounts}
\title{Normalise barcode counts}
\usage{
normaliseCounts(dgeObject, method = "CPM", threshold = 0)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{method}{method of normalization. One of `CPM`, `TMM`, `TMMwsp`, `upperquartile` or `RLE` (string). See `edgeR::calcNormFactors()`.}

\item{threshold}{Threshold to apply to counts before normalisation (integer). Default = `0`.}
}
\value{
Returns a DGEList object containing normalised barcode counts
}
\description{
Normalise barcode counts within a DGEList object
}
\examples{
data(test.dge)
test.dge.normalized <- normaliseCounts(test.dge, method = "CPM", threshold = 0)
head(test.dge.normalized)

}
