% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBoxplot.R
\name{plotBarcodeBoxplot}
\alias{plotBarcodeBoxplot}
\title{Boxplot of selected barcodes}
\usage{
plotBarcodeBoxplot(
  dgeObject,
  barcodes = NULL,
  group = NULL,
  conditions = NULL,
  trans = NULL,
  point = FALSE,
  violin = FALSE,
  returnData = FALSE,
  normalizeMethod = "CPM"
)
}
\arguments{
\item{dgeObject}{DGEList object with normalized barcode counts.}

\item{barcodes}{Barcodes to be plotted (vector of strings).}

\item{group}{Optional, column name in sample metadata to facet data on (string).}

\item{conditions}{Optional, specific levels of group to be plotted (vector of strings).}

\item{trans}{Optional, transformation of y-axis, e.g. `log10` (string).}

\item{point}{Whether to include points (boolean). Default = `FALSE`.}

\item{violin}{Whether to include violin plots in addition to box plots (boolean). Default = `FALSE`.}

\item{returnData}{Whether to return data instead of plot (boolean). Default = `FALSE`.}

\item{normalizeMethod}{Method for normalizing counts (string). One of `CPM`, `TMM`, `TMMwsp`, `upperquartile`, `RLE` or `NULL` for no normalization. See `edgeR::calcNormFactors()`. Default = `CPM`.}
}
\value{
Returns a boxplot
}
\description{
Plots boxplot of counts for selected barcodes from DGEList object.
}
\examples{
data(test.dge)
plotBarcodeBoxplot(test.dge, barcodes = "BC_190202", group = "Treatment",
conditions = c("Vehicle", "Low_dose", "High_dose"))
}
