% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBoxplot.R
\name{plotBarcodeBoxplot}
\alias{plotBarcodeBoxplot}
\title{Boxplot of selected barcodes}
\usage{
plotBarcodeBoxplot(dge.obj, barcodes, condition = NULL)
}
\arguments{
\item{dge.obj}{DGE object containing counts and sample information}

\item{barcodes}{vector of barcodes to be plotted}

\item{condition}{specific metadata condition to be plotted}
}
\value{
Returns a boxplot
}
\description{
Takes a dataframe of barcode counts,
select specific barcodes in the dataframe
and boxplot their counts
}
\details{
plotBCBoxplot

Boxplot of selected barcodes
}
\examples{
data(test.dge)
barcodes <- sample(rownames(test.dge), 10)
plotBarcodeBoxplot(test.dge, barcodes)
plotBarcodeBoxplot(test.dge, barcodes, condition = c("T0", "T0b"))
}
