% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeCumSum.R
\name{plotBarcodeCumSum}
\alias{plotBarcodeCumSum}
\title{Barcode cumulative sum plot of a list of samples against and ordered by sample1}
\usage{
plotBarcodeCumSum(counts, sample1 = NULL, samples = NULL)
}
\arguments{
\item{counts}{Dataframe containing cpm counts}

\item{sample1}{sample to compare others against}

\item{samples}{vector of sample names to be plotted against sample1}
}
\value{
Returns a plot of the cumulative sum of read counts per barcode
}
\description{
Takes a dataframe of the barcode cpm counts, calculate the relative abundance of each barcode in each sample.
Then, barcodes of all samples are ordered for sample1 in decreasing order and the cumulative sum is calculated for
each sample. Cumulative sum is then plotted against sample1.
}
\details{
plotBarcodeCumSum

Plot the cumulative sum of barcode cpm counts for a list of samples against and ordered by sample1.
}
\examples{
data(test.counts)
plotBarcodeCumSum(counts = test.counts, sample1 = 'T0-1', samples = c('T0-1','T0b-1'))
}
