% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeDist.R
\name{plotBarcodeDistance}
\alias{plotBarcodeDistance}
\title{Sample to sample distances}
\usage{
plotBarcodeDistance(
  counts,
  method = "euclidean",
  upper = T,
  clustered = T,
  name = "Sample Distances"
)
}
\arguments{
\item{counts}{matrix/dataframe containing raw or normalised counts}

\item{method}{distance metric to use. Refer to stats::dist for available options}

\item{upper}{Logical. plot only the upper half of the matrix}

\item{clustered}{Logical. cluster rows and columns of matrix}

\item{name}{title of plot}
}
\value{
Returns a heatmap of sample distances using desired clustering
}
\description{
Plot sample distances between barcode sets / samples
}
\examples{
test.mat <- matrix(rnorm(10*10,mean=0,sd=2), 10, 10)
plotBarcodeDistance(test.mat)
}
