% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeHeatmap.R
\name{plotBarcodeHeatmap}
\alias{plotBarcodeHeatmap}
\title{Plot barcode heatmap}
\usage{
plotBarcodeHeatmap(
  dgeObject,
  topN = 5,
  name = "CPM",
  showBarcodes = FALSE,
  group = NULL,
  colAnnot = NULL,
  discrete = FALSE,
  discreteThreshold = 1,
  labelTopBarcodes = TRUE
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{topN}{Number of top barcodes per sample to show (integer). Default = `5`.}

\item{name}{String of the heatmap scale name (string). Default = `CPM`.}

\item{showBarcodes}{Show barcode names on the heatmap (boolean). Default = `FALSE`.}

\item{group}{Optional, column name in sample metadata to group samples by (string).}

\item{colAnnot}{List of vectors assigning colours to each level of metadata.}

\item{discrete}{Show presence/absence of barcodes instead of abundance (boolean). Default = `FALSE`.}

\item{discreteThreshold}{Threshold for presence of barcode in samples (decimal). Default = `1`.}

\item{labelTopBarcodes}{Label top barcodes within a sample (boolean). Default = `TRUE`.}
}
\value{
Returns a heatmap
}
\description{
Takes a DGEList with barcode counts, selects the n most abundant barcodes per sample and plots a heatmap.
Stars indicate the most abundant barcodes in the respective sample.
}
\examples{
data(test.dge)
plotBarcodeHeatmap(test.dge)
}
