% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeHeatmap.R
\name{plotBarcodeHeatmap}
\alias{plotBarcodeHeatmap}
\title{Plot barcode heatmap}
\usage{
plotBarcodeHeatmap(counts, N, name = "CPM", show_bc = FALSE)
}
\arguments{
\item{counts}{Dataframe containing cpm counts}

\item{N}{String of sample 1 name (used to ranked barcode abundance)}

\item{name}{String of the heatmap scale name}

\item{show_bc}{Boolean to show barcode names on the heatmap}
}
\value{
Returns a heatmap
}
\description{
Takes a dataframe of the barcode counts, and selects the N most abundant barcodes per sample.
Most abundant barcodes are gathered in a list and a heatmap for those barcodes is plotted for all samples.
Stars indicate the most abundant barcodes per sample.
}
\details{
plotBarcodeHeatmap

Plot a heatmap of the N most abundant barcodes per sample
}
\examples{
data(test.counts)
plotBarcodeHeatmap(test.counts, 10,'Counts',FALSE)
}
