% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeHistogram.R
\name{plotBarcodeHistogram}
\alias{plotBarcodeHistogram}
\title{plotBarcodeHistogram}
\usage{
plotBarcodeHistogram(
  counts.obj,
  sample = NULL,
  top = 10,
  name = "Barcode Histogram"
)
}
\arguments{
\item{counts.obj}{dataframe containing raw counts of barcodes}

\item{sample}{desired name of sample to order all barcodes against}

\item{top}{top n barcodes in sample to color, all other barcodes are shown in grey}

\item{name}{desired plot title}
}
\value{
Returns a barcode histogram plot of barcodes represented by proportion of total pool
}
\description{
Generate stacked barcode plots showing proportion from raw count object.
}
\examples{
data(test.counts)
plotBarcodeHistogram(test.counts, sample = "T0-1", top = 10, name = "Barcode Histogram")
}
