% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodePCA.R
\name{plotBarcodePCA}
\alias{plotBarcodePCA}
\title{Barcode PCA plot}
\usage{
plotBarcodePCA(
  object,
  intgroup = "condition",
  col = "group",
  ntop = 500,
  returnData = FALSE,
  batch = NULL
)
}
\arguments{
\item{object}{DGEList or dataframe containing normalised barcode counts}

\item{intgroup}{vector of grouping variables of interest}

\item{col}{color group for plot, must be one of intgroup}

\item{ntop}{number of top most variable genes to be used in PCA calculation}

\item{returnData}{Logical. return a data.frame of PCA calculation?}

\item{batch}{metadata category indicating source of batch effects. Will be used in limma removeBatchEffect function prior to PCA}
}
\value{
Returns a plot of the first two principal components in the dataset
}
\description{
Takes an edgeR DGEList object contining barcode data or a dataframe of the barcode cpm counts and performs PCA.
Then, plots the first two dimensions.
Modified from DESeq2 package plotPCA function
}
\details{
plotBarcodePCA

Plot the first two principal components of the barcode dataset
}
\examples{
data(test.dge)
plotBarcodePCA(test.dge, intgroup = "group", ntop = 500, returnData = FALSE, batch = NULL)
}
