% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeRegression.R
\name{plotBarcodeRegression}
\alias{plotBarcodeRegression}
\title{plotBarcodeRegression}
\usage{
plotBarcodeRegression(
  dge,
  samp1 = NULL,
  samp2 = NULL,
  title = NULL,
  trendline = T,
  trans = NULL,
  rug = F
)
}
\arguments{
\item{dge}{DGEList object containing grouping variable to fit linear model}

\item{samp1}{name of sample 1. must be one of colnames(dge)}

\item{samp2}{name of sample 2. must be one of colnames(dge)}

\item{title}{desired name of output plot}

\item{trendline}{Logical. Include linear trendline using stat_smooth()}

\item{trans}{From ggplot2. For continuous scales, the name of a transformation object or the object itself.}

\item{rug}{Logical. Include geom_rug density information on the axes?}
}
\value{
Returns a scatterplot of the two samples
}
\description{
Generate a linear regression scatterplot for two sets of sample counts.
}
\examples{
data(test.dge)
plotBarcodeRegression(dge = test.dge, samp1 = "T0-1", samp2 = "T0-2")
}
