% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellBARtabQC.R
\name{plotBarcodesPerCell}
\alias{plotBarcodesPerCell}
\title{plotBarcodesPerCell}
\usage{
plotBarcodesPerCell(
  counts,
  fraction = TRUE,
  aggregated = FALSE,
  notDetected = "",
  sep = ";"
)
}
\arguments{
\item{counts}{Dataframe with barcodes and UMI counts per cell}

\item{fraction}{Boolean, whether to plot fraction or number of cells}

\item{aggregated}{Counts were aggregated per cell (boolean). Default = `FALSE`.}

\item{notDetected}{Optional, string representing no detected barcode. NA is always treated as not detected.}

\item{sep}{Separating character used for aggregation (string). Default = `;`.}
}
\value{
Returns a plot
}
\description{
Plot number of detected barcodes per cell.
}
