% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeTimeseries.R
\name{plotBarcodeTimeseries}
\alias{plotBarcodeTimeseries}
\title{plotBarcodeTimeseries}
\usage{
plotBarcodeTimeseries(counts.obj, name = "", seed = 5)
}
\arguments{
\item{counts.obj}{dataframe containing raw counts of barcodes}

\item{name}{desired plot title}

\item{seed}{RNG seed}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate proportional timeseries plot from raw / normalised barcode count object.
}
\examples{
data(test.counts)
plotBarcodeTimeseries(test.counts[,1:4], name = "Proportional Timeseries Plot", seed = 5)
}
