% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsinClusters.R
\name{plotCellsInClusters}
\alias{plotCellsInClusters}
\title{Plot cells per ident class}
\usage{
plotCellsInClusters(
  sc.obj,
  group = NULL,
  factor = NULL,
  idents = NULL,
  plot.pct = T,
  plot = T
)
}
\arguments{
\item{sc.obj}{single cell object in Seurat or SingleCellExperiment format containing idents and group metadata}

\item{group}{a column of metadata}

\item{factor}{a level of group to plot per level of idents}

\item{idents}{a column of metadata defining the cluster identities of each cell. default = "seurat_clusters"}

\item{plot.pct}{Logical. Plot percentages (TRUE) or raw numbers (FALSE). default = TRUE}

\item{plot}{Logical. Plot or return raw data. default = TRUE}
}
\value{
Returns a histogram or underlying plot data
}
\description{
Takes a single cell object, a grouping variable, a factor within the group to test, 
and an ident class (i.e. clusters). Plots percentage or raw number of cells within factor
present per level of ident (i.e. per cluster)
}
\details{
Plot number or percentage of a group per level of ident in a single
cell object
}
