% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterEnrichment.R
\name{plotClusterEnrichment}
\alias{plotClusterEnrichment}
\title{Hypergeometric test for cluster enrichment}
\usage{
plotClusterEnrichment(
  sc.obj = NULL,
  group = NULL,
  factor = NULL,
  idents = "seurat_clusters",
  threshold = 0.01,
  order = TRUE,
  plot = TRUE
)
}
\arguments{
\item{sc.obj}{single cell object in Seurat or SingleCellExperiment format containing idents and group metadata}

\item{group}{a column of metadata}

\item{factor}{a level of group to test for enrichment per level of ident}

\item{idents}{a column of metadata defining the cluster identities of each cell. default = "seurat_clusters"}

\item{threshold}{P-value threshold for phyper test. default = 0.01}

\item{order}{Logical. Order levels of idents in plot by -log10 P-value}

\item{plot}{Logical. Plot hypergeometric test results or return plot data.}
}
\value{
Returns a histogram of hypergeometric test results or underlying plot data
}
\description{
Takes a single cell object, a grouping variable, a factor within the group to test, 
and an ident class (i.e. clusters). Per level of ident performs hypergeometric testing
for enrichment of group factor
}
\details{
plotClusterEnrichment

Hypergeometric test for enrichment of specific classes of cells in 
single cell RNA-seq clusters
}
