% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDetectedBarcodes.R
\name{plotDetectedBarcodes}
\alias{plotDetectedBarcodes}
\title{plotDetectedBarcodes}
\usage{
plotDetectedBarcodes(counts, percentile = 0.95, plot = T)
}
\arguments{
\item{counts}{DGEList or dataframe containing raw or normalised barcode counts}

\item{percentile}{desired percentile value. 95th percentile by default}

\item{plot}{Logical. plot data instead of returning counts table.}
}
\value{
Returns a histogram plot of the number of detected barcodes per sample.
}
\description{
Plot the total number of barcodes detected in a sample
}
\examples{
data(test.dge)
plotDetectedBarcodes(test.dge, percentile = .95)
plotDetectedBarcodes(test.dge, plot = FALSE)
}
