% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivIndexes.R
\name{plotDivIndexes}
\alias{plotDivIndexes}
\title{Plot diversity indices across barcode samples}
\usage{
plotDivIndexes(
  counts,
  div = NULL,
  metric = "shannon",
  type = "bar",
  group = NULL
)
}
\arguments{
\item{counts}{dataframe of raw or normalised counts with samples as columns and observations/barcodes as rows}

\item{div}{dataframe of diversity metrics as columns and samples as rows}

\item{metric}{diversity metric to plot. one of "shannon", "simpson", "invsimpson" or "gini"}

\item{type}{plot as bar graph or point}

\item{group}{a character vector of containing grouping information. Must be equal the number of columns of counts. Can pass a metadata column from DGEList object.}
}
\value{
Returns a plot of calculated diversity index per sample
}
\description{
Plots a dataframe of diversity indices per sample
}
\details{
plotDivIndexes

Plots common diversity indices per sample
}
\examples{
plotDivIndexes(counts = test.dge$counts)
}
