% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrderedBubble.R
\name{plotOrderedBubble}
\alias{plotOrderedBubble}
\title{plotOrderedBubble}
\usage{
plotOrderedBubble(
  dgeObject,
  title = "Proportional Bubble Plot",
  orderSample = NULL,
  group = NULL,
  displaySamples = NULL,
  displayBarcodes = NULL,
  proportionCutoff = 10,
  colorDominant = FALSE,
  filterCutoff = NULL,
  labelBarcodes = TRUE,
  legend = TRUE,
  pseudoCount = FALSE
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{title}{Plot title (string). Default = `Proportional Bubble Plot`.}

\item{orderSample}{Name of sample to order by (string).}

\item{group}{Optional, column name in sample metadata to group samples by (string).}

\item{displaySamples}{Optional, vector of samples to display - keep the order of vector.}

\item{displayBarcodes}{Optional, vector of barcodes to display.}

\item{proportionCutoff}{barcodes represented at a percentage within any sample above this threshold will be labelled (decimal). Default = `10`.}

\item{colorDominant}{Only color clones with frequency above `proportionCutoff` and others grey (boolean). Default = `FALSE`.}

\item{filterCutoff}{Barcodes below this threshold in `orderSample` will be filtered in all samples (boolean). Default = `TRUE`.}

\item{labelBarcodes}{Label barcodes with a proportion larger than proportionCutoff in any sample (boolean). Default = `TRUE`.}

\item{legend}{Show a legend of bubble sizes (boolean). Default = `TRUE`.}

\item{pseudoCount}{Whether to add a pseudo count of 1 to all counts to display barcodes absent in T0 (boolean). Requires counts to be normalized. Default = `FALSE`.}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate ordered proportional bubbleplots from raw count object with barcodes labelled above a specified threshold
}
\examples{
data(test.dge)
plotOrderedBubble(test.dge, orderSample = "T0-1", filterCutoff = 0.001, group = "Treatment")
}
