% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrderedBubble.R
\name{plotOrderedBubble}
\alias{plotOrderedBubble}
\title{plotOrderedBubble}
\usage{
plotOrderedBubble(
  counts.obj,
  labels = T,
  name = "Proportional Bubble Plot",
  orderSample = NULL,
  proportion.cutoff = 10,
  colorDominant = FALSE
)
}
\arguments{
\item{counts.obj}{dataframe containing raw counts of barcodes. assumes barcodes as rownames.}

\item{labels}{logical. Print barcode labels}

\item{name}{desired plot title}

\item{orderSample}{name of sample to order by}

\item{proportion.cutoff}{barcodes represented at a percentage within any sample above this threshold will be labelled}

\item{colorDominant}{only color clones with frequency above proportion.cutoff. Others colored grey}
}
\value{
Returns a bubbleplot of barcodes represented by proportion of total pool
}
\description{
Generate ordered proportional bubbleplots from raw count object with barcodes labelled above a specified threshold
}
\examples{
plotOrderedBubble(test.dge$counts, orderSample = "T0-1")
}
