% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadCounts.R
\name{plotReadCounts}
\alias{plotReadCounts}
\title{Plot sample read counts}
\usage{
plotReadCounts(
  counts,
  group = NULL,
  log10 = FALSE,
  legend = TRUE,
  order = TRUE
)
}
\arguments{
\item{counts}{data.frame of barcode count x sample}

\item{group}{a character vector of containing grouping information. Must be equal the number of columns of counts. Can pass a metadata column from DGEList object.}

\item{log10}{Boolean. log10 transform output?}

\item{legend}{Boolean. Include legend?}

\item{order}{Boolean. Order samples by group?}
}
\value{
Returns a plot of the read counts per column (sample) in a data frame
}
\description{
Simple plot of total read counts per sample
}
\examples{
data(test.counts)
plotReadCounts(test.dge$counts, group = test.dge$samples$Group)

}
