% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.counts}
\alias{test.counts}
\title{Test barcode counts}
\format{
A data frame with 100 rows and 4 variables:
\describe{
  \item{T0-1}{sample T0, replicate 1}
  \item{T0-2}{sample T0, replicate 2}
  \item{T0b-1}{sample T0b, replicate 1}
  \item{T0b-2}{sample T0b, replicate 2}
  ...
}
}
\usage{
test.counts
}
\description{
A dataset containing normalised count information for 100 hypothetical barcodes
in 4 hypothetical samples
}
\keyword{datasets}
