% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.dge}
\alias{test.dge}
\title{Test barcode DGEList object}
\format{
A DGEList object containing 100 rows and 4 variables:
}
\usage{
test.dge
}
\description{
A DGEList object containing raw count information for 100 hypothetical barcodes
in 4 hypothetical samples
}
\keyword{datasets}
