% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdCounts.R
\name{thresholdCounts}
\alias{thresholdCounts}
\title{Threshold counts}
\usage{
thresholdCounts(
  dgeObject,
  threshold = 10,
  type = "absolute",
  minSamples = 1,
  plot = FALSE,
  group = NULL,
  order = TRUE
)
}
\arguments{
\item{dgeObject}{DGEList object with barcode counts.}

\item{threshold}{The threshold to use. If type = "relative", must a float between 0 & 1. Default = `10`}

\item{type}{Threshold type to use (string). Must be one of "absolute" or "relative". Default = `absolute`.}

\item{minSamples}{Minimum number of samples a barcode must meet threshold to remain in dataset (integer). Default = `1`.}

\item{plot}{Return plot instead of filtered data (boolean). Default = `FALSE`.}

\item{group}{Optional, column name in sample metadata to color samples by (string).}

\item{order}{Order samples by group (boolean). Default = `TRUE`.}
}
\value{
Returns a filtered DGEList object.
}
\description{
Filter barcodes meeting a given absolute (total read count) or relative (proportion based) abundance level
Optionally plot number of barcodes detected using this threshold in each sample.
}
\examples{
thresholdCounts(test.dge, type = "absolute", threshold = 10)
thresholdCounts(test.dge, type = "absolute", threshold = 10, plot = TRUE)
}
