% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdCounts.R
\name{thresholdCounts}
\alias{thresholdCounts}
\title{Threshold counts}
\usage{
thresholdCounts(df, threshold = 20, plot = FALSE)
}
\arguments{
\item{df}{Dataframe to be thresholded.}

\item{threshold}{The threshold to use. Rows with count below this will be removed.}

\item{plot}{Logical. Draw plots of dataset?}
}
\value{
Returns a thresholded data-frame
}
\description{
Threshold dataframe to a given level and return number of barcodes meeting threshold in each
sample.
}
\examples{
thresholdCounts(test.counts, threshold = 20, plot = FALSE)
}
